refactor: separate service and controller - moved service and controller class into respective packages - split service into poinManagementService and PoinHistoryManagementService
