+++ src/main/kotlin/io/hhplus/tdd/controller/dto/ChargePointRequest.kt
@@ -1,8 +1,8 @@
-package io.hhplus.tdd.point.dto
+package io.hhplus.tdd.controller.dto
 
 import jakarta.validation.constraints.Min
 
-data class UsePointRequest(
+data class ChargePointRequest(
     @field:Min(value = 0, message = "amount must be greater than or equal to zero")
     val amount: Long = 0
 )
