+++ src/main/kotlin/io/hhplus/tdd/controller/dto/UsePointRequest.kt
@@ -1,8 +1,8 @@
-package io.hhplus.tdd.point.dto
+package io.hhplus.tdd.controller.dto
 
 import jakarta.validation.constraints.Min
 
-data class ChargePointRequest(
+data class UsePointRequest(
     @field:Min(value = 0, message = "amount must be greater than or equal to zero")
     val amount: Long = 0
 )
